backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
pydantic-settings>=2.1.0
httpx>=0.25.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-google-genai>=2.0.0

